Rename components to index
