+++ src/app/App.tsx
@@ -20,11 +20,11 @@ import useAppManager from "./hooks/useAppManager";
 import useWindowResize from "./hooks/useWindowResize";
 
 // components
-import InformationScreen from "./components/layout/screens/information-screen/InformationScreen";
-import Debug from "./components/layout/debug/Debug";
-import Main from "./components/layout/main/Main";
-import Sidebar from "./components/layout/sidebar/Sidebar";
-import Minimize from "./components/common/buttons/minimize/Minimize";
+import InformationScreen from "./components/layout/screens/information-screen";
+import Debug from "./components/layout/debug";
+import Main from "./components/layout/main";
+import Sidebar from "./components/layout/sidebar";
+import Minimize from "./components/common/buttons/minimize";
 
 import "./index.less";
