+++ src/app/components/common/buttons/minimize/index.tsx
