+++ src/app/components/common/list/index.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react";
 
 // components
-import ListItem, { IListItem, IItemComponentProps } from "./list-item/ListItem";
+import ListItem, { IListItem, IItemComponentProps } from "./list-item";
 
 // styles
 import "./index.less";
