+++ src/app/components/common/progress-bar/index.tsx
